Added runat.sh - universal top-level runner.

+++ tests/hpcc/run.sh
@@ -12,41 +12,32 @@ if [ -f "${HPCHUB_PLATFORM}" ]; then
   . ${HPCHUB_PLATFORM}
 fi
 
-if [ "${HPCHUB_REPORT}" = "" ]; then
-  report_to=../report.time.txt
-else
-  report_to=${HPCHUB_REPORT}
-fi
-
-
-Tprev=`date +%s.%N`
-
-function LogStep {
-   ret=$?
-   if [ "$ret" -gt "0" ]; then
-      echo "Step $StepCntr failed"
-      exit $ret
-   else
-      StepCntr=$((StepCntr+1))
-   fi
-   local T0=`date +%s.%N`
-   dT=`echo "$T0 - $Tprev" | bc -l`
-   echo -ne "${p}\t$1\t$dT\t$T0\n" >> $report_to
-   Tprev=$T0
-}
-
-echo -ne "Protein\tStep\tdT\tTimeStamp\n" > $report_to
-
 cd hpcc-${hpcc_version}
 
 cp ../hpccinf.txt ./
 
-cp ../machinefile ./
+if [ -f "${HPCHUB_MACHINEFILE}" ];then
+  cp "${HPCHUB_MACHINEFILE}" ./machinefile
+else
+  cp ../machinefile ./
+fi
 
-LogStep Start 
+LogStep hpcc-noargs Start 
+
+rm hpccoutf.txt 2>/dev/null  
 
 ${HPCHUB_MPIRUN} `pwd`/hpcc
 
-LogStep hpcc-all
+${HPCHUB_MPIWAIT}
+
+LogStep hpcc-noargs hpcc-all
+
+NMPI=`grep -A 10 "Begin of MPIRandomAccess section." hpccoutf.txt | grep "GUP" | head -n 1 | awk '{print $1;};'` 
+
+LogStep hpcc-noargs MPIRandomAccess $NMPI
+
+NStar=`grep -A 10 "Begin of StarRandomAccess section." hpccoutf.txt | grep "GUP" | head -n 1 | awk '{print $1;};'` 
+
+LogStep hpcc-noargs MPIRandomAccess $NStar
 
 cd ..
